refactor to layout => YT page
